+++ src/app/pathogen-notification/components/favorites-list/favorites-list.component.ts
@@ -37,11 +37,12 @@ export class FavoritesListComponent extends FieldType implements OnInit, OnDestr
 
   favorites: WritableSignal<CodeDisplay[]> = signal([]);
   protected readonly getDesignationValueIfAvailable = getDesignationValueIfAvailable;
-  private readonly isNonNominalNotification7_3: boolean = false;
+  private readonly isNotification7_3: boolean = false;
 
   constructor() {
     super();
-    this.isNonNominalNotification7_3 = this.pathogenNotificationComponent.isNonNominalNotification7_3();
+    this.isNotification7_3 =
+      this.pathogenNotificationComponent.isNonNominalNotification7_3() || this.pathogenNotificationComponent.isAnonymousNotification7_3();
     this.loadFavorites();
   }
 
@@ -53,7 +54,7 @@ export class FavoritesListComponent extends FieldType implements OnInit, OnDestr
   }
 
   loadFavorites(): void {
-    const favorites = this.pathogenNotificationStorageService.getFavorites(this.isNonNominalNotification7_3) || [];
+    const favorites = this.pathogenNotificationStorageService.getFavorites(this.isNotification7_3) || [];
     this.favorites.update(() => favorites);
   }
 
@@ -69,7 +70,7 @@ export class FavoritesListComponent extends FieldType implements OnInit, OnDestr
 
   removePathogenFromFavorites(pathogen: CodeDisplay): void {
     const updatedFavorites = this.favorites().filter(fav => fav.code !== pathogen.code);
-    this.pathogenNotificationStorageService.updateFavorites(updatedFavorites, this.isNonNominalNotification7_3);
+    this.pathogenNotificationStorageService.updateFavorites(updatedFavorites, this.isNotification7_3);
   }
 
   ngOnDestroy(): void {