+++ src/test/integration/nonnominal.spec.ts
@@ -21,6 +21,7 @@ import { HarnessLoader } from '@angular/cdk/testing';
 import { MockedComponentFixture } from 'ng-mocks';
 import { switchToPage } from '../shared/test-utils';
 import { NotificationType } from '../../app/pathogen-notification/common/routing-helper';
+import { NotifiedPersonDisclaimer } from '../../app/pathogen-notification/utils/disclaimer-texts';
 
 describe('Pathogen - Nonnominal Integration Tests', () => {
   let component: PathogenNotificationComponent;
@@ -64,5 +65,10 @@ describe('Pathogen - Nonnominal Integration Tests', () => {
 
     expect(federalStateSelect).withContext('federalStateSelect should not be present but was found').toBeNull();
     expect(pathogenDisplaySelect).withContext('pathogenDisplaySelect could not be found').toBeTruthy();
+  });
+
+  it('should show the default disclaimer text on notified person page', async () => {
+    await switchToPage(3, fixture);
+    expect(fixture.nativeElement.textContent).toContain(NotifiedPersonDisclaimer.DEFAULT_DISCLAIMER);
   });
 });